requirements.txt
# Telegram Bot Framework
aiogram==3.3.0
aiohttp==3.9.1

# Database
aiosqlite==0.19.0

# Web Scraping & Browser Automation
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Image Processing
Pillow==10.1.0

# Async Support
asyncio-throttle==1.0.2

# Configuration
python-dotenv==1.0.0

# HTTP Client & Data Processing
httpx==0.25.2
python-dateutil==2.8.2
